v1.2 update

+++ reddit_scraper_v1.2.py
@@ -1,4 +1,4 @@
-#reddit_scraper_v1.1
+#reddit_scraper_v1.2
 print("Reddit scraper")
 
 import praw
@@ -7,7 +7,7 @@ import os
 
 reddit = praw.Reddit('main_account', user_agent='main_account user agent') #this logs into a reddit account specified in praw.ini
 
-subreddit_input = input("Enter subreddit: ")
+subreddit_input = input("Enter subreddit: /r/")
 image_number = int(input("Enter the number of submissions to parse (limit 1000): "))
 
 subreddit = reddit.subreddit(subreddit_input)
@@ -18,8 +18,11 @@ for submission in subreddit.new(limit=image_number):
     file_extension = submission.url[len(submission.url) -4:]
     print("file_extension: " + file_extension)
     if file_extension == ".jpg" or file_extension == "png" or file_extension == ".bmp" or file_extension == "jpeg": #if url is a direct image link
-        file_name = wget.download(submission.url,)    
-        os.rename(file_name, "downloads/" + submission.title + " .jpg") 
-        print("\n")
+        if os.path.exists("downloads/" + submission.title + " .jpg") == False:
+            file_name = wget.download(submission.url,)
+            os.rename(file_name, "downloads/" + submission.title + " .jpg") 
+            print("\n")
+        else:
+            print("File already downloaded.\n")
     else:
         print("Submission is not a direct image link.")
